Migrate tests that use redirect_url
